msm: camera: common: Add level in priority list for shutdown
Currently sensor and csiphy has the same priority in shutdown
sequence. This is sometime calls sensor shutdown before csiphy
and with the AON usecase it is failing with CPAS related operation.
This change introducing the new enum entry for the Medium_low, which
make sure the csiphy shutdown gets call before the sensor shutdown
in order to make the operation in align with cross modules.

CRs-Fixed: 3084672
Change-Id: I40f495151392f96ac89caf8df852ed8326b79d9b
Signed-off-by: Jigarkumar Zala <quic_jzala@quicinc.com>