msm: camera: isp: Reorganize IFE bus ver3 for v880
Update the correct WM idx for ALSC client for v880. The
change also removes dependency on comp group type for
buf done irq masking. The order is not maintained in HW
so breaking that assumption in SW, and making the buf done
irq bit mask as an array for comp groups.

CRs-Fixed: 3254772
Change-Id: I7ac1dcf4150761b2f6f124e389a5e5b1ea67bd83
Signed-off-by: Karthik Anantha Ram <quic_kartanan@quicinc.com>