msm: camera: isp: Memory optimization for out resources
Vfe out resources memory allocation is done at the time of
probe based on max out ports supported. Considering the
number of contexts and the size of each node, the total
size comes around 16KB whereas in normal use cases, this
much memory is way too large. For upcoming multi context
targets, this number will be multifold, roughly around
48KB.
This commit changes the memory design to maintain the
out resources. With this commit, at the time of probe,
around 400 bytes will be allocated and rest of the
memory will be allocated at the time of acquire and freed
at the time of release. Only the required memory will be
allocated.

CRs-Fixed: 3321317
Change-Id: Ibb069962736a90bcf89ecfd0f54f983094be942a
Signed-off-by: Gaurav Jindal <quic_gjindal@quicinc.com>