msm: camera: isp: Modify reset type on CSID stream off
Currently prior to issuing a SW reset to CSID, we are
ensuring the input to CSID rx is cut off. On any fatal
rx error, the driver tries to issue a SW global reset
without halting the bus. If bus is not IDLE, it could
lead to an invalid domain ID getting propagated leading to
a global fault. To avoid this use path reset on CSID stream off,
global reset to be used on use-case boundaries when we are
certain that the bus is IDLE.

CRs-Fixed: 3420832
Change-Id: Id3beb42963357c881d073c258b7357360797b4bf
Signed-off-by: Karthik Anantha Ram <quic_kartanan@quicinc.com>