msm: camera: isp: Enhance frame timing IRQ debugging
In VFE top, currently it is assumed that frame timing
IRQs will be part of status 1 register. Remove that
assumption to support v980 hw as well. Also, when
frame irq debug is enabled in ife hw manager, don't
trigger callback, as we just want to see if frame
reached IFE/TFE or not.

CRs-Fixed: 3321317
Change-Id: I544727d4bec628b9041531015f24bf73e8ae2be1
Signed-off-by: Mukund Madhusudan Atre <quic_matre@quicinc.com>