msm: camera: icp: Modify the session dump content of ICP
Do minimum memory length check when do session dump of ICP,
in case of the memory size is not enough. We can't copy the
content in icp hw mgr device info in to the target memory
because ldar parser doesn't know the data format of it, so
ldar parser unableto  parse it. For the clock info, we need
to add device name info to it in order to distinguish the
owner of the clock info.

CRs-Fixed: 3554419
Change-Id: Id616c27054947fc09816ec1aadb98a882e52ca91
Signed-off-by: mingpan <quic_mingpan@quicinc.com>