msm: camera: common: Add generic usage of start axi paths for cpas clients
Currently, during cpas start, cpas clients specify a certain axi path
that is suitable to vote for as a start vote. But, due to changes
in axi paths frequently in targets, the start path might not be valid.
Add a generic start path which acts as an indication to pick the first
valid path found for that target. Add usage of this start path only for
dynamic update RT clients.

CRs-Fixed: 3321317
Change-Id: I0a73487950bed80d282af3cdbbef7060776f4e1d
Signed-off-by: Mukund Madhusudan Atre <quic_matre@quicinc.com>