msm: camera: isp: Add support for leading context irqs in mc cases
Currently, in mc cases, we listen to composite irqs for sof and
rup done. But, for certain single context clients, the frame might
be done before the composite irqs come. Add support to listen to ipp0
path irqs which is guaranteed to have leading context, so that isp
state machine can function accordingly.

CRs-Fixed: 3321317
Change-Id: I6a1af5a9994dacbf81c8ea65bfcb25c730bf1dbc
Signed-off-by: Mukund Madhusudan Atre <quic_matre@quicinc.com>