msm: camera: isp: Add support for master slave cfg in aeb mode
Currently, primary secondary, or master slave config is only
programmed in shdr and rdi lcr cases. With csid driving the isp
state machine as the targets progress, we need rup-aup sync
with rdi0 as well. Add support to configure master slave config
in aeb case to facilitate this sync.

CRs-Fixed: 3649197
Change-Id: Ia7b8a38e0f83e1f403feb141d974df057e01d519
Signed-off-by: Mukund Madhusudan Atre <quic_matre@quicinc.com>