msm: camera: isp: Fix unbalanced dependent unregister for csid irq
Currently, we look for next zero bit in dependent bitmap of the primary
controller to register a secondary controller, which does not check if
the bitmap has a bit set already for that controller. Add validation of
dependent register and unregister calls balance in irq controller.
Add missing unregister dependent calls from csid for top2 controller.
Also, fix the controller pointer passed for mc irq unregister in error
cases.

CRs-Fixed: 3321317
Change-Id: I9367308b0ffa127e804657bc448eb3cab523ad7f
Signed-off-by: Mukund Madhusudan Atre <quic_matre@quicinc.com>