msm: camera: common: Disable flex array checks in UBSAN for camera module
LTS kernel 6.4 onwards there is upstream compiler change is added to
check the array bound, which is called flex array check.
Though we are allocating the enough memory for array, but still it will
look for specific flex array(uint32_t foo[]) required size need to be
calculated separately. We have array with size [1] is there manu places,
until we are able to solve all the places we will continue this with
change.

CRs-Fixed: 3656124
Change-Id: I95aa30bb847f22befe105d857fb1f063ae23f0ce
Signed-off-by: Soumen Ghosh <quic_soumeng@quicinc.com>