msm: camera: cpas: Change sw client voting method for camnoc axi clk
Currently, camnoc axi clk consolidates a graph tree to obtain the value
needed to vote. This change incorporates another value to this process.
The applied clk rate for each sw client is now passed onto cpas
interface and is used for voting axi clk. The clk will be set to the
max of the passed client value and the value obtained from the
graph tree.

CRs-Fixed: 3653079
Change-Id: Idf0c24adf07c1191da08225feeb4501db42bb060
Signed-off-by: Atiya Kailany <quic_akailany@quicinc.com>