msm: camera: common: Add gpio_for_vmrm_purpose for vmrm
Due to cci is using pinctrl, but it does not define gpio info
and vmrm is checking the gpios parameter to get the device use
gpio info, so need to add gpio for cci. And due to cci is using
pinctrl, so there is not need to define gpio-req-tbl-num for cci,
so cci just has gpios without gpio-req-tbl-num. This commit add
gpio_for_vmrm_purpose to identify the scene.

CRs-Fixed: 3683555
Change-Id: I78afd856d8eaf700c1cfa10e774713c735f595a9
Signed-off-by: zhuo <quic_zhuo@quicinc.com>