msm: camera: csid: Add support for handling CRC error
This change turns CAM_IFE_CSID_RX_ERROR_CRC to part
fatal error and flags it as fatal one when the number
of such error reaches the threshold that is calculated
by 1% of the frame width. CRC error counter is reset
when getting the SOF timestamp.

This change also turns CRC error into recoverable one.
Internal recovery is first triggered when the number of
such error reaches the threshold.

CRs-Fixed: 3428500
Change-Id: I5f5caf1b2394b415fd7c20da29ff1649cd08d554
Signed-off-by: Haochen Yang <quic_haocyang@quicinc.com>