msm: camera: csiphy: Expand CSIPHY Onthego support
Currently, onthego sysfs module only supports
configuration for one PHY at a time, and the
PHY cannot be selected, the configuration
committed will be applied to whichever PHY that is
streamed on after the settings is committed.

This change adds support to program multiple
settings for each PHY, and support for specifying
a particular PHY for which the setting is to be
applied.

CRs-Fixed: 3681877
Change-Id: Ia9555491dbfe40c83837897ac20662d9491dd82d
Signed-off-by: Li Sha Lim <quic_lishlim@quicinc.com>