msm: camera: isp: Check all the out ports in CSID acquire
In the secure camera use cases with both Pixel and RDI ports
enabled from the CSID, while acquiring the pixel resource, we
are updating the secure flag of that CSID Eventually when we
acquire the RDI resource on the same CSID, we are checking the
secure mode of the first output port for the input port, which
typically, is non secure stats, with the secure mode of the
CSID that's previously set. Thus, the comparison fails and we
go on to acquire the IFE Lite. Acquiring a TFE and an IFE Lite
in the same session is not supported in the secure camera use cases.

CRs-Fixed: 3718783
Change-Id: I42e02c5c8c04d88ea23deadd0ed771edd6d29d6a
Signed-off-by: Vijay Kumar Tumati <quic_vtumati@quicinc.com>